requirements.txt
# Core Dependencies
click>=8.0.0
PyYAML>=6.0.0
python-dotenv>=1.0.0

# Machine Learning and Optimization
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Database
SQLAlchemy>=2.0.0

# Async Support
aiohttp>=3.9.0
asyncio>=3.4.3

# Dashboard
dash>=2.14.0
plotly>=5.18.0
dash-bootstrap-components>=1.5.0

# API
fastapi>=0.105.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
email-validator>=2.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
pytest-mock>=3.14.0
httpx>=0.27.0  # For testing FastAPI endpoints
hypothesis>=6.122.4  # For property-based testing